fix: type분리와 함수 수정
