+++ App.tsx
@@ -1,4 +1,3 @@
-// App.tsx
 import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
@@ -7,12 +6,6 @@ import { RootStackParamList } from "./src/types/navigation";
 import LoginScreen from "./src/screens/Login/LoginScreen";
 import KakaoLoginWebview from "./src/screens/Login/KakaoLoginWebview";
 import LoginSuccessScreen from "./src/screens/Login/LoginSuccess";
-
-/** ─────────────────────────────
- *  네비게이터 파라미터 타입
- *  (Zustand 전역 스토어 사용 → 스택 파라미터 불필요)
- *  ────────────────────────────*/
-
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
