+++ src/api/KakaoLogin/kakao_api.ts
@@ -1,16 +1,8 @@
-// src/api/kakao_api.ts
-// Kakao OAuth helper (axios + TypeScript)
-// =============================================================
-// - Interfaces `KakaoToken`, `KakaoProfile`
-// - Strict type annotations on functions
-// - Easily migratable from previous JS version
-
 import axios from "axios";
-import qs from "qs";
+import qs from "qs"; // 작동은 잘 되는데 왜 빨간 밑줄 에러가 뜰까..?
+import { KakaoToken } from "../../types/kakao_api";
+import { KakaoProfile } from "../../types/kakao_api";
 
-// ────────────────────────────────────────────────────────────────
-// Kakao OAuth constants
-// ────────────────────────────────────────────────────────────────
 export const REST_KEY = "ce11a8c09ff42d519d2257d412297dc7";
 export const REDIRECT_URI = "https://dummy.emobin.app/oauth";
 
@@ -20,38 +12,12 @@ export const AUTH_URL =
     client_id: REST_KEY,
     redirect_uri: REDIRECT_URI,
     response_type: "code",
-    scope: "profile_nickname", // 이메일 scope 제외
+    scope: "profile_nickname",
   }).toString();
 
-// ────────────────────────────────────────────────────────────────
-// Type definitions
-// ────────────────────────────────────────────────────────────────
 
-export interface KakaoToken {
-  access_token: string;
-  refresh_token: string;
-  expires_in: number;
-  token_type: string;
-  scope: string;
-}
 
-export interface KakaoProfile {
-  id: string; // Kakao user id (문자열)
-  profile: {
-    nickname: string;
-    profile_image_url?: string;
-    thumbnail_image_url?: string;
-  };
-  email?: string;
-}
 
-// ────────────────────────────────────────────────────────────────
-// API helpers
-// ────────────────────────────────────────────────────────────────
-
-/**
- * 교환된 code 값을 OAuth 토큰으로 변환합니다.
- */
 export async function exchangeKakaoToken(code: string): Promise<KakaoToken> {
   try {
     const { data } = await axios.post<KakaoToken>(
@@ -78,10 +44,7 @@ export async function exchangeKakaoToken(code: string): Promise<KakaoToken> {
   }
 }
 
-/**
- * 액세스 토큰으로 카카오 사용자 프로필 조회
- * 반환 값은 사용 편의성을 위해 핵심 필드만 정규화합니다.
- */
+
 export async function fetchKakaoProfile(
   accessToken: string
 ): Promise<KakaoProfile> {
@@ -110,9 +73,6 @@ export async function fetchKakaoProfile(
   }
 }
 
-// ────────────────────────────────────────────────────────────────
-// Optional: axios instance for reuse
-// ────────────────────────────────────────────────────────────────
 
 export const kakaoApi = axios.create({
   baseURL: "https://kapi.kakao.com",
