+++ src/components/Button.style.ts
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  kakaoButton: {
+    width: 310,
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    backgroundColor: "#FEE500",
+    paddingVertical: 14,
+    paddingHorizontal: 20,
+    borderRadius: 24,
+  },
+  kakaoIcon: {
+    width: 18,
+    height: 18,
+    marginRight: 8,
+  },
+  kakaoText: {
+    fontSize: 16,
+    fontWeight: "bold",
+    color: "#000",
+  },
+});
