+++ src/screens/Login/LoginSuccess.tsx
@@ -1,10 +1,3 @@
-// src/screens/LoginSuccess.tsx
-// 로그인 성공 후 표시되는 Splash-like 화면
-// -------------------------------------------------------------
-// React Native 0.65+ 부터 `BackHandler.removeEventListener()` 는 deprecated.
-// addEventListener 가 돌려주는 subscription 객체의 `remove()` 를 호출해야 합니다.
-// 이 버전에 맞게 수정했습니다.
-
 import React, { useEffect } from "react";
 import {
   View,
@@ -14,23 +7,16 @@ import {
   Platform,
 } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import type { NativeStackScreenProps } from "@react-navigation/native-stack";
-import { RootStackParamList } from "../../types/navigation";
 import { styles } from "./LoginSuccess.style";
+import { PropsSuccess } from "../../types/login"
 
-// 로고 이미지 (프로젝트 구조에 맞게 경로 조정)
 const logo = require("../../../assets/logo.png");
 
-type Props = NativeStackScreenProps<RootStackParamList, "LoginSuccess">;
 
-export default function LoginSuccessScreen({}: Props) {
-  // Android 하드웨어 back 버튼 비활성화
+const LoginSuccessScreen: React.FC<PropsSuccess> = () => {
   useEffect(() => {
     if (Platform.OS === "android") {
-      const subscription = BackHandler.addEventListener(
-        "hardwareBackPress",
-        () => true // true 반환 → 이벤트 소비, 뒤로가기 무시
-      );
+      const subscription = BackHandler.addEventListener("hardwareBackPress", () => true); // 뒤로가기 막기
       return () => subscription.remove();
     }
   }, []);
@@ -44,4 +30,6 @@ export default function LoginSuccessScreen({}: Props) {
       <Text style={styles.footer}>The Synthetics</Text>
     </SafeAreaView>
   );
-}
+};
+
+export default LoginSuccessScreen;
