+++ src/stores/authStore.ts
@@ -1,32 +1,7 @@
-// src/stores/authStore.ts
-// zustand + persist 로 간단한 인증 상태 관리
-
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
 
-// ────────────────────────────────────────────────────────────────
-// KakaoProfile 타입 (API helper 와 동일한 구조)
-// ────────────────────────────────────────────────────────────────
-
-export interface KakaoProfile {
-  id: string; // Kakao user id (문자열)
-  profile: {
-    nickname: string;
-    profile_image_url?: string;
-    thumbnail_image_url?: string;
-  };
-  email?: string;
-}
-
-// ────────────────────────────────────────────────────────────────
-// Zustand 상태 정의
-// ────────────────────────────────────────────────────────────────
-
-interface AuthState {
-  profile: KakaoProfile | null;
-  setProfile: (p: KakaoProfile) => void;
-  logout: () => void;
-}
+import { AuthState } from "../types/auth";
 
 export const useAuthStore = create<AuthState>()(
   persist(
@@ -36,8 +11,7 @@ export const useAuthStore = create<AuthState>()(
       logout: () => set({ profile: null }),
     }),
     {
-      name: "auth", // AsyncStorage key
-      // storage: createJSONStorage(() => SecureStore) ← Expo SecureStore 사용 시
+      name: "auth",
     }
   )
 );
