+++ src/types/button.ts
@@ -1,0 +1,7 @@
+import { GestureResponderEvent } from "react-native";
+
+export interface ButtonProps {
+  text: string;
+  onPress: (event: GestureResponderEvent) => void;
+  iconSource: any;
+}
