+++ src/types/login.ts
@@ -1,0 +1,6 @@
+import type { NativeStackScreenProps } from "@react-navigation/native-stack";
+import { RootStackParamList } from "../types/navigation";
+
+export type PropsSuccess = NativeStackScreenProps<RootStackParamList, "LoginSuccess">;
+
+export type Props = NativeStackScreenProps<RootStackParamList, "Login">;
